+++ src/drivers/generators/BaseCacheGenerator.php
@@ -8,6 +8,7 @@ namespace putyourlightson\blitz\drivers\generators;
 use Amp\MultiReasonException;
 use Craft;
 use craft\base\SavableComponent;
+use craft\helpers\Console;
 use putyourlightson\blitz\Blitz;
 use putyourlightson\blitz\events\RefreshCacheEvent;
 use putyourlightson\blitz\helpers\CacheGeneratorHelper;
@@ -15,6 +16,7 @@ use putyourlightson\blitz\helpers\SiteUriHelper;
 use putyourlightson\blitz\models\SiteUriModel;
 use putyourlightson\blitz\services\CacheRequestService;
 use Throwable;
+use yii\helpers\BaseConsole;
 
 /**
  * @property-read array $siteOptions
@@ -72,9 +74,14 @@ abstract class BaseCacheGenerator extends SavableComponent implements CacheGener
     public const GENERATE_ACTION_ROUTE = 'blitz/generator/generate';
 
     /**
+     * @var bool Whether to output verbose messages.
+     */
+    public bool $verbose = false;
+
+    /**
      * @inheritdoc
      */
-    public function generateUris(array $siteUris, callable $setProgressHandler = null, bool $queue = true): void
+    public function generateUris(array $siteUris, callable $setProgressHandler = null, bool $queue = true, bool $verbose = false): void
     {
         $event = new RefreshCacheEvent(['siteUris' => $siteUris]);
         $this->trigger(self::EVENT_BEFORE_GENERATE_CACHE, $event);
@@ -108,7 +115,7 @@ abstract class BaseCacheGenerator extends SavableComponent implements CacheGener
     /**
      * @inheritdoc
      */
-    public function generateSite(int $siteId, callable $setProgressHandler = null, bool $queue = true): void
+    public function generateSite(int $siteId, callable $setProgressHandler = null, bool $queue = true, bool $verbose = false): void
     {
         // Get custom site URIs for the provided site only
         $groupedSiteUris = SiteUriHelper::getSiteUrisGroupedBySite(Blitz::$plugin->settings->getCustomSiteUris());
@@ -119,13 +126,13 @@ abstract class BaseCacheGenerator extends SavableComponent implements CacheGener
             $customSiteUris
         );
 
-        $this->generateUris($siteUris, $setProgressHandler, $queue);
+        $this->generateUris($siteUris, $setProgressHandler, $queue, $verbose);
     }
 
     /**
      * @inheritdoc
      */
-    public function generateAll(callable $setProgressHandler = null, bool $queue = true): void
+    public function generateAll(callable $setProgressHandler = null, bool $queue = true, bool $verbose = false): void
     {
         $event = new RefreshCacheEvent();
         $this->trigger(self::EVENT_BEFORE_GENERATE_ALL_CACHE, $event);
@@ -139,7 +146,7 @@ abstract class BaseCacheGenerator extends SavableComponent implements CacheGener
             Blitz::$plugin->settings->getCustomSiteUris()
         );
 
-        $this->generateUris($siteUris, $setProgressHandler, $queue);
+        $this->generateUris($siteUris, $setProgressHandler, $queue, $verbose);
 
         if ($this->hasEventHandlers(self::EVENT_AFTER_GENERATE_ALL_CACHE)) {
             $this->trigger(self::EVENT_AFTER_GENERATE_ALL_CACHE, $event);
@@ -234,5 +241,19 @@ abstract class BaseCacheGenerator extends SavableComponent implements CacheGener
     protected function isPageUrl(string $url): bool
     {
         return !str_contains($url, CacheRequestService::CACHED_INCLUDE_PATH . '?action=' . CacheRequestService::CACHED_INCLUDE_ACTION);
+    }
+
+    protected function outputVerbose(string $url, bool $success = true): void
+    {
+        if (Craft::$app->getRequest()->getIsConsoleRequest() && $this->verbose) {
+            $tokenParam = Craft::$app->getConfig()->getGeneral()->tokenParam;
+            $url = preg_replace('/[?&]' . $tokenParam . '.*/', '', $url);
+
+            if ($success) {
+                Console::stdout($url . PHP_EOL);
+            } else {
+                Console::stdout($url . PHP_EOL, BaseConsole::FG_RED);
+            }
+        }
     }
 }